'Metro' style for 'Frame' control (part 2 FIX)
The problem was because MetroWindowHelpers class is internal, those
extension members needed to be in a public class instead.
